requirements.txt
deepagents==0.4.0
python-dotenv==1.2.1
tavily==1.1.0
